rename all instances of data availability to celestia
